The resource_range_ip* classes were recently changed to use rpki.POW.IPAddress as the underlying address representation.  POW.X509.getPrefixes() returns the prefix as an rpki.POW.IPAddress, so return it directly without a cast to long.
closes #352


svn path=/trunk/; revision=4883

+++ rpkid/rpki/rcynic.py
@@ -109,10 +109,10 @@ class rcynic_roa(rcynic_object):
     v4, v6 = self.obj.get_POW().getPrefixes()
     if v4:
       self.prefix_sets.append(rpki.resource_set.roa_prefix_set_ipv4([
-        rpki.resource_set.roa_prefix_ipv4(long(p[0]), p[1], p[2]) for p in v4]))
+        rpki.resource_set.roa_prefix_ipv4(p[0], p[1], p[2]) for p in v4]))
     if v6:
       self.prefix_sets.append(rpki.resource_set.roa_prefix_set_ipv6([
-        rpki.resource_set.roa_prefix_ipv6(long(p[0]), p[1], p[2]) for p in v6]))
+        rpki.resource_set.roa_prefix_ipv6(p[0], p[1], p[2]) for p in v6]))
     self.ee = rpki.x509.X509(POW = self.obj.get_POW().certs()[0])
     self.notBefore = self.ee.getNotBefore()
     self.notAfter = self.ee.getNotAfter()
